implementing file transfer progress

+++ cloud-common/src/main/java/files/service/FileTransferProgressService.java
@@ -5,7 +5,7 @@ import java.nio.file.Path;
 public interface FileTransferProgressService {
 
     void add(Path path, long size);
-    void increment(Path path, long transferred);
+    void increment(Path path, int transferred);
     float progress(Path path);
     void remove(Path path);
 }
